Nov 20, 2023, 10:12 PM

+++ aliases_func.c
@@ -1,10 +1,11 @@
 #include "main.h"
+#include <stddef.h>
 
 /* Helper function to handle alias commands */
 int handle_alias_command(char *buffer, alias_t **aliases)
 {
 	char *alias_prefix = "alias ";
-	int alias_prefix_len = _strlen(alias_prefix);
+	size_t alias_prefix_len = _strlen(alias_prefix);
 
 	if (strncmp(buffer, alias_prefix, alias_prefix_len) == 0)
 	{
